slight tweaks to how the scenes are selected
I think a direct reference to the scene itself is a bit more robust. If we change the scene name we won't have to change in these scripts where we reference it. Also, if we decide to connect to a new scene entirely we can just pick it instead of typing out its name.